UCX-Backend: Initial UCX Backend Implementation
* In the src/utils directory, there is a new header and class implementation that abstracts away the trickier parts of interacting with UCX and provides simpler functions.
* Also in src/utils there is a test executable that tests the functionality of this ucx_utils class. This test runs successfully with both DRAM and VRAM.
* In src/nixl_nw_backends, we have a tested implementation of the ucx_backend that implements the NIXL backend_engine class
* In the new test directory, there is a sample application that directly interacts with the ucx_backend to test functionality.
* Some header files and src/nixl files have been updated to reflect some API changes we have discussed.

Merge Request: adv-dev-team/nixl!7

+++ src/nixl_descriptors.cpp
@@ -36,8 +36,8 @@ basic_desc::basic_desc(const basic_desc& desc){
 
 bool basic_desc::covers (const basic_desc& query) const {
     if (dev_id == query.dev_id) {
-        if (((uint64_t) addr <= (uint64_t) query.addr) &&
-            ((uint64_t) addr + len >= (uint64_t) query.addr + query.len))
+        if (( addr <=  query.addr) &&
+            ( addr + len >= query.addr + query.len))
             return true;
     }
     return false;
